Adjust synchronizing flags, debug output if block relays are being skipped

+++ src/TAO/Ledger/chainstate.cpp
@@ -72,11 +72,11 @@ namespace TAO
 
             /* Check if there's been a new block. */
             static memory::atomic<uint1024_t> hashLast;
-            static std::atomic<uint32_t> nLastTime;
+            static std::atomic<uint64_t> nLastTime;
             if(hashBestChain.load() != hashLast.load())
             {
                 hashLast = hashBestChain.load();
-                nLastTime = static_cast<uint32_t>(runtime::unifiedtimestamp());
+                nLastTime = runtime::unifiedtimestamp();
             }
 
             /* Special testnet rule. */
